requirements.txt
Flask==2.0.3
Werkzeug==2.0.3

google-cloud-dialogflow==2.14.1
python-dotenv==0.19.1
requests==2.26.0
